Abort for GC thread panic (#221)
The code is mainly from `mmrk-ruby`. We hijack the Rust panic hook to
make sure that if the GC threads panic, the process will be aborted. The
current behavior is that when a GC thread panics, the process hangs and
wait for that GC thread to finish its work.

This should be ported to `master`.